backend/requirements.txt
# Core Framework
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-multipart==0.0.22
python-dotenv==1.0.1

# Database
motor==3.3.1
pymongo==4.5.0

# Authentication
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0

# MikroTik API
RouterOS-api==0.17.0

# HTTP Client
httpx==0.27.0
requests==2.31.0

# SSH (untuk backup via SSH export)
paramiko==3.4.0
cryptography>=41.0.0

# Validation
pydantic==2.6.0
python-dateutil==2.8.2

# Server-Sent Events (real-time dashboard)
sse-starlette==2.1.3

# RADIUS Server Integration
pyrad==2.5.4
netaddr>=0.10.1

# SNMP — pysnmp-lextudio: fork resmi yang fix Python 3.11 asyncio.coroutine issue
# Drop-in replacement, API sama persis (getCmd, walkCmd, dll tetap sama)
pysnmp-lextudio
pyasn1>=0.4.8

# Push Notifications FCM
firebase-admin==6.4.0
